requirements.txt
psycopg2-binary==2.9.9
mysql-connector-python==8.3.0
apache-airflow-providers-airbyte
apache-airflow-providers-openlineage>=1.8.0
protobuf==4.21.12
urllib3==1.26.16